Create reference-style dark landing page
